fix: refactored helpers
