+++ src/shared/modules/helpers.ts
@@ -3,7 +3,7 @@ import get from 'lodash/get';
 import invert from 'lodash/invert';
 import isEmpty from 'lodash/isEmpty';
 import pick from 'lodash/pick';
-import {META_KEY, CURRENT_VERSION} from '../constants';
+import {META_KEY, CURRENT_VERSION, ACTION_PARAM_PREFIX} from '../constants';
 import {
     PluginBase,
     ConfigItem,
@@ -16,6 +16,7 @@ import {
     ConfigItemWithTabs,
     Config,
     QueueItem,
+    ItemStateAndParams,
 } from '../types';
 
 function getNormalizedPlugins(plugins: PluginBase[]) {
@@ -284,4 +285,77 @@ export function deleteFromQueue(data: ChangeQueueArg): StateAndParamsMetaData {
         ...meta,
         queue: meta.queue.slice(0, -1),
     };
+}
+
+export function pickActionParamsFromParams(
+    params: ItemStateAndParams['params'],
+    returnWithPrefix?: boolean,
+) {
+    if (!params || isEmpty(params)) {
+        return {};
+    }
+
+    const actionParams: StringParams = {};
+    for (const [key, val] of Object.entries(params)) {
+        // starts with actionParams prefix (from'_ap_')
+        if (key.startsWith(ACTION_PARAM_PREFIX)) {
+            const paramName = returnWithPrefix ? key : key.substr(ACTION_PARAM_PREFIX.length);
+            actionParams[paramName] = val;
+        }
+    }
+    return actionParams;
+}
+
+/**
+ * public function for getting params from object without actionParams
+ * @param params
+ */
+export function pickExceptActionParamsFromParams(params: ItemStateAndParams['params']) {
+    if (!params || isEmpty(params)) {
+        return {};
+    }
+
+    const onlyParams: StringParams = {};
+    for (const [key, val] of Object.entries(params)) {
+        if (!key.startsWith(ACTION_PARAM_PREFIX)) {
+            onlyParams[key] = val;
+        }
+    }
+    return onlyParams;
+}
+
+/**
+ * public function for transforming object to actionParams format
+ * @param params
+ */
+export function transformParamsToActionParams(params: ItemStateAndParams['params']) {
+    if (!params || isEmpty(params)) {
+        return {};
+    }
+
+    const actionParams: StringParams = {};
+    for (const [key, val] of Object.entries(params)) {
+        actionParams[`${ACTION_PARAM_PREFIX}${key}`] = val;
+    }
+    return actionParams;
+}
+
+/**
+ * check if object contains actionParams
+ * @param conf
+ */
+function hasActionParam(conf?: StringParams): boolean {
+    return Boolean(Object.keys(conf || {}).find((key) => key.startsWith(ACTION_PARAM_PREFIX)));
+}
+
+/**
+ * check if ItemStateAndParams object has actionParams in params or state field
+ * @param stateAndParams
+ */
+export function hasActionParams(stateAndParams: ItemStateAndParams) {
+    if (!stateAndParams || isEmpty(stateAndParams)) {
+        return false;
+    }
+
+    return hasActionParam(stateAndParams.params) || hasActionParam(stateAndParams.state);
 }
